refactor(theme): rename styles/cn.ts to styles/index.ts
