+++ packages/theme/src/components/ArchiveCalendar.astro
@@ -1,5 +1,5 @@
 ---
-import { cn } from "../styles/cn"
+import { cn } from "../styles"
 
 interface ArchiveDayItem {
   day: number
