+++ packages/theme/src/components/ArchiveLinks.astro
@@ -1,5 +1,5 @@
 ---
-import { cn } from "../styles/cn"
+import { cn } from "../styles"
 
 interface ArchiveLinkItem {
   label: string
