+++ packages/theme/src/components/Aside.astro
@@ -5,7 +5,7 @@ import PostList from "./PostList.astro"
 import SearchForm from "./SearchForm.astro"
 import Heatmap from "./Heatmap.astro"
 
-import { cn } from "../styles/cn"
+import { cn } from "../styles"
 import { t, themeConfig } from "../runtime/config"
 import { getNotes, getPosts, getPinnedPosts, getSeries } from "../content/content"
