+++ packages/theme/src/components/Heatmap.astro
@@ -2,7 +2,7 @@
 import { countBy } from "lodash-es"
 import { eachDayOfInterval, startOfDay, startOfWeek, subWeeks } from "date-fns"
 
-import { cn } from "../styles/cn"
+import { cn } from "../styles"
 import { t } from "../runtime/config"
 import { getNotes, getPosts } from "../content/content"
 import { getArchiveBounds, getDayNotesPath } from "../content/archive"
