+++ packages/theme/src/components/NoteList.astro
@@ -3,7 +3,7 @@ import Section from "./Section.astro"
 import Pagination from "./Pagination.astro"
 import NoteContent from "./NoteContent.astro"
 
-import { cn } from "../styles/cn"
+import { cn } from "../styles"
 import { t } from "../runtime/config"
 import { groupNotesByDate } from "../content/content"
 import { getDayNotesPath } from "../content/archive"
