+++ packages/theme/src/components/NoteSearchResults.astro
@@ -2,7 +2,7 @@
 import Pagination from "./Pagination.astro"
 import NoteContent from "./NoteContent.astro"
 
-import { cn } from "../styles/cn"
+import { cn } from "../styles"
 import { t } from "../runtime/config"
 
 import type { CollectionEntry } from "astro:content"
