+++ packages/theme/src/components/PostList.astro
@@ -1,7 +1,7 @@
 ---
 import PostTitle from "./PostTitle.astro"
 
-import { cn } from "../styles/cn"
+import { cn } from "../styles"
 import { t } from "../runtime/config"
 
 import type { CollectionEntry } from "astro:content"
