+++ packages/theme/src/components/PostTitle.astro
@@ -1,5 +1,5 @@
 ---
-import { cn } from "../styles/cn"
+import { cn } from "../styles"
 import { getSeriesById } from "../content/content"
 
 import type { CollectionEntry } from "astro:content"
