+++ packages/theme/src/components/SearchForm.astro
@@ -1,5 +1,5 @@
 ---
-import { cn } from "../styles/cn"
+import { cn } from "../styles"
 import { t } from "../runtime/config"
 
 interface Props {
