+++ packages/theme/src/components/Section.astro
@@ -1,5 +1,5 @@
 ---
-import { cn } from "../styles/cn"
+import { cn } from "../styles"
 
 interface Props {
   title?: string
