+++ packages/theme/src/components/SiteCard.astro
@@ -4,7 +4,7 @@ import { Image } from "astro:assets"
 import logoMesh from "../assets/logo-mesh.png"
 import logoMuse from "../assets/logo-muse.png"
 
-import { cn } from "../styles/cn"
+import { cn } from "../styles"
 
 import type { ImageMetadata } from "astro"
