+++ packages/theme/src/components/TableOfContents.astro
@@ -3,7 +3,7 @@ import { render } from "astro:content"
 
 import Section from "./Section.astro"
 
-import { cn } from "../styles/cn"
+import { cn } from "../styles"
 import { t } from "../runtime/config"
 
 import type { CollectionEntry } from "astro:content"
