+++ packages/theme/src/index.ts
@@ -1,3 +1,3 @@
-export { cn } from "./styles/cn"
+export { cn } from "./styles"
 
 export type { PrimaryColor } from "./config"
