+++ packages/theme/src/routes/notes/[...slug].astro
@@ -2,7 +2,7 @@
 import Layout from "../../components/Layout.astro"
 import NoteContent from "../../components/NoteContent.astro"
 
-import { cn } from "../../styles/cn"
+import { cn } from "../../styles"
 import { t } from "../../runtime/config"
 import { getAdjacentNotes, getNoteById, getPostById } from "../../content/content"
