+++ packages/theme/src/routes/posts/[...slug].astro
@@ -6,7 +6,7 @@ import PostTitle from "../../components/PostTitle.astro"
 import TableOfContents from "../../components/TableOfContents.astro"
 import PostList from "../../components/PostList.astro"
 
-import { cn } from "../../styles/cn"
+import { cn } from "../../styles"
 import { t } from "../../runtime/config"
 import { getAdjacentPosts, getPosts, getPostById, getSeriesById, filterPostsInSeries } from "../../content/content"
